Consolidate UI template tests
Reduce test runtime aand easier to maintain.
